+++ src/ShapeCrawler/Tables/TopBottomMargin.cs
@@ -6,7 +6,7 @@ namespace ShapeCrawler.Tables;
 
 internal readonly ref struct TopBottomMargin(Int32Value? emus)
 {
-    const float DefaultTopAndBottomMargin = 3.69f; // ~0.13 cm
+    const decimal DefaultTopAndBottomMargin = 3.69m; // ~0.13 cm
 
-    public float Value => emus is null ? DefaultTopAndBottomMargin : new Emus(emus).AsPoints();
+    public decimal Value => emus is null ? DefaultTopAndBottomMargin : new Emus(emus).AsPoints();
 }
