+++ src/ShapeCrawler/Texts/LeftRightMargin.cs
@@ -5,7 +5,7 @@ namespace ShapeCrawler.Texts;
 
 internal readonly ref struct LeftRightMargin(Int32Value? emus)
 {
-    const float DefaultLeftAndRightMarginPoints = 7.09f; // ~0.25 cm
+    const decimal DefaultLeftAndRightMarginPoints = 7.09m; // ~0.25 cm
 
-    public float Value => emus is null ? DefaultLeftAndRightMarginPoints : new Emus(emus).AsPoints();
+    public decimal Value => emus is null ? DefaultLeftAndRightMarginPoints : new Emus(emus).AsPoints();
 }
